requirements.txt
aniso8601==9.0.1
asgiref==3.3.1
astroid==2.4.2
asttokens==2.0.5
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.1.3
colorama==0.4.4
Django==3.1.3
django-smtp-ssl==1.0
djangorestframework==3.13.1
executing==0.7.0
Flask==2.1.2
Flask-RESTful==0.3.9
icecream==2.1.1
idna==3.2
importlib-metadata==4.11.4
isort==5.6.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.3
MarkupSafe==2.1.1
mccabe==0.6.1
Pillow==8.4.0
Pygments==2.9.0
pylint==2.6.0
pytz==2020.4
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
